Add SNMP Polling of Instances
Adds polling of all instances for QPS. Filtering of instances
still remains to be done.

Instances are polled on the management network defined in
ceilometer.conf. This must be a network routable from the host
address space, or you must use floating IPs. The default is to
use the floating IP from the instance's port on that network.