Move to central polling agent
Since generally the collection of the NIOS SNMP data will be done
over floating IPs, there is no benefit to polling on the compute
nodes. Thus, the entry point is changed to poll from the central
agent, removing the need to install this on the compute nodes.

Additionally, bump the version and thin the requirements (remove
the requirement for the ceilometer git repo).